requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
supabase>=2.0.0
python-dotenv>=1.0.0
certifi>=2023.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
